backend/requirements.txt
jaseci>=0.5.0
jaseci-serv>=0.5.0
python-dotenv>=0.19.0
requests>=2.26.0
flask>=2.0.0
flask-cors>=3.0.10
sqlalchemy>=1.4.0
pydantic>=1.8.0
openai>=0.27.0
anthropic>=0.3.0
ollama>=0.1.0
pytest>=6.0.0
pytest-cov>=2.12.0
python-dateutil>=2.8.0
